Split Field.fromDouble to enable reuse.
